# be/docs/openapi/cmms-api.yaml
openapi: 3.1.0
info:
  title: CMMS API
  version: 1.0.0
  description: >
    Computerized Maintenance Management System REST API.
    Modules: Auth, Dashboard, Equipment, Work Orders, Schedule, Technicians,
    Spare Parts, Reports, Alerts, Settings. Roles: Admin, Manager, Technician.
servers:
  - url: http://localhost:5000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/work-orders:
    get:
      tags: [WorkOrders]
      summary: List work orders (paginated, filterable)
      parameters:
        - { name: type, in: query, schema: { $ref: '#/components/schemas/WorkOrderType' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/WorkOrderStatus' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/WorkOrderPriority' } }
        - { name: technicianId, in: query, schema: { type: string, format: uuid } }
        - { name: equipmentId, in: query, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/SortBy' }
        - { $ref: '#/components/parameters/SortDir' }
      responses:
        '200':
          description: Paginated work orders
          content:
            application/json:
              examples:
                detail:
                  externalValue: ./examples/work-orders.json#/workOrderDetail
    post:
      tags: [WorkOrders]
      summary: Create work order (Admin, Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkOrder' }
            examples:
              corrective:
                summary: Corrective WO - motor overheating
                externalValue: ./examples/work-orders.json#/createCorrective
              preventive:
                summary: Preventive WO - monthly inspection
                externalValue: ./examples/work-orders.json#/createPreventive
      responses:
        '201':
          description: Created
          content:
            application/json:
              examples:
                created:
                  externalValue: ./examples/work-orders.json#/workOrderDetail
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/work-orders/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [WorkOrders]
      summary: Get work order by id
      responses:
        '200':
          description: Work order detail
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/work-orders.json#/workOrderDetail
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [WorkOrders]
      summary: Update work order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkOrder' }
      responses:
        '200': { description: Updated }
    delete:
      tags: [WorkOrders]
      summary: Soft-delete work order (Admin, Manager)
      responses:
        '204': { description: Deleted }
  /api/work-orders/{id}/status:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    patch:
      tags: [WorkOrders]
      summary: Change work order status (state machine enforced)
      description: >
        Allowed transitions: Draft->Assigned, Assigned->InProgress|Draft,
        InProgress->OnHold|Completed, OnHold->InProgress, Completed->Closed.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderStatusChange' }
            examples:
              start:
                externalValue: ./examples/work-orders.json#/statusChange
      responses:
        '200':
          description: Status changed
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/work-orders.json#/workOrderDetail
        '409':
          description: Invalid transition
          content:
            application/json:
              examples:
                invalid:
                  externalValue: ./examples/errors.json#/businessRule
...
  /api/schedules:
    get:
      tags: [Schedule]
      summary: List maintenance schedules
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paginated schedules }
    post:
      tags: [Schedule]
      summary: Create schedule (Admin, Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSchedule' }
            examples:
              monthly:
                externalValue: ./examples/schedule.json#/createMonthly
              byMeter:
                externalValue: ./examples/schedule.json#/createByMeter
      responses:
        '201': { description: Created }
  /api/schedules/calendar:
    get:
      tags: [Schedule]
      summary: Calendar events between two dates
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Calendar events
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/schedule.json#/calendarEvents
  /api/schedules/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    put:
      tags: [Schedule]
      summary: Update schedule (Admin, Manager)
      responses:
        '200': { description: Updated }
    delete:
      tags: [Schedule]
      summary: Delete schedule (Admin, Manager)
      responses:
        '204': { description: Deleted }
  /api/technicians:
    get:
      tags: [Technicians]
      summary: List technicians with workload (Admin, Manager)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Paginated technicians
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/technicians.json#/list
  /api/technicians/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Technicians]
      summary: Technician profile
      responses:
        '200': { description: Technician }
  /api/technicians/{id}/workload:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Technicians]
      summary: Technician workload breakdown
      responses:
        '200':
          description: Workload
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/technicians.json#/workload
  /api/spare-parts:
    get:
      tags: [SpareParts]
      summary: List spare parts (Admin, Manager)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paginated spare parts }
    post:
      tags: [SpareParts]
      summary: Create spare part (Admin, Manager)
      responses:
        '201': { description: Created }
  /api/spare-parts/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [SpareParts]
      summary: Get spare part
      responses:
        '200': { description: Spare part }
    put:
      tags: [SpareParts]
      summary: Update spare part (Admin, Manager)
      responses:
        '200': { description: Updated }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    SortBy:
      name: sortBy
      in: query
      schema: { type: string }
    SortDir:
      name: sortDir
      in: query
      schema: { type: string, enum: [asc, desc] }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            default:
              externalValue: ./examples/errors.json#/notFound
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            default:
              externalValue: ./examples/errors.json#/validation
  schemas:
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
        details:
          type: array
          items: { type: string }
      required: [error, code]
...
    WorkOrderType:
      type: string
      enum: [Corrective, Preventive, Inspection]
    WorkOrderStatus:
      type: string
      enum: [Draft, Assigned, InProgress, OnHold, Completed, Closed]
    WorkOrderPriority:
      type: string
      enum: [Low, Medium, High, Critical]
    ScheduleFrequency:
      type: string
      enum: [Daily, Weekly, Monthly, ByMeter]
...
    ChecklistItemInput:
      type: object
      properties:
        description: { type: string }
        isDone: { type: boolean }
    CreateWorkOrder:
      type: object
      required: [type, priority, equipmentId, description]
      properties:
        type: { $ref: '#/components/schemas/WorkOrderType' }
        priority: { $ref: '#/components/schemas/WorkOrderPriority' }
        equipmentId: { type: string, format: uuid }
        assignedTechnicianId: { type: string, format: uuid, nullable: true }
        description: { type: string }
        deadline: { type: string, format: date-time, nullable: true }
        checklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItemInput' }
    WorkOrderStatusChange:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/WorkOrderStatus' }
    CreateSchedule:
      type: object
      required: [equipmentId, title, frequency, intervalValue, nextDueDate]
      properties:
        equipmentId: { type: string, format: uuid }
        title: { type: string }
        frequency: { $ref: '#/components/schemas/ScheduleFrequency' }
        intervalValue: { type: integer }
        meterThreshold: { type: number, nullable: true }
        nextDueDate: { type: string, format: date }
        isActive: { type: boolean }
...